refactor: moved route handlers to seperate module
